+++ db/seeds.rb
@@ -1,9 +1,7 @@
 # frozen_string_literal: true
 
-# This file should contain all the record creation needed to seed the database with its default values.
-# The data can then be loaded with the bin/rails db:seed command (or created alongside the database with db:setup).
-#
-# Examples:
-#
-#   movies = Movie.create([{ name: "Star Wars" }, { name: "Lord of the Rings" }])
-#   Character.create(name: "Luke", movie: movies.first)
+app = FactoryBot.create :application
+user = FactoryBot.create :user
+
+puts "\e[32mApplication created: #{app.name}\e[0m \n\033[1mUID:\033[0m #{app.uid} \n\033[1mSECRET:\033[0m #{app.secret}\n\n"
+puts "\e[32mUser created: #{user.id}\e[0m \n\033[1memail:\033[0m #{user.email} \n\033[1mpassword:\033[0m Pa$$w0rd\n\n"